more TC5 APIs backported.

+++ src/main/java/train/common/entity/rollingStock/EntityLocoSteamForneyRed.java
@@ -161,10 +161,10 @@ public class EntityLocoSteamForneyRed extends SteamTrain {
 	}
 
 	@Override
-	public int transportTopSpeed() { return super.transportTopSpeed()*(ridingEntity instanceof EntityPlayerMP &&(((EntityPlayerMP) ridingEntity).getDisplayName().equals("EternalBlueFlame") || ((EntityPlayerMP) ridingEntity).getDisplayName().equals("minecarftmano9"))?3:1);}
+	public float transportTopSpeed() { return super.transportTopSpeed()*(ridingEntity instanceof EntityPlayerMP &&(((EntityPlayerMP) ridingEntity).getDisplayName().equals("EternalBlueFlame") || ((EntityPlayerMP) ridingEntity).getDisplayName().equals("minecarftmano9"))?3:1);}
 
 	@Override
-	public int transportMetricHorsePower() {
+	public float transportMetricHorsePower() {
 		return super.transportMetricHorsePower()*(ridingEntity instanceof EntityPlayerMP &&(((EntityPlayerMP) ridingEntity).getDisplayName().equals("EternalBlueFlame") || ((EntityPlayerMP) ridingEntity).getDisplayName().equals("minecarftmano9"))?3:1);
 	}
 }
